requirements.txt
python-telegram-bot==20.7
google-generativeai==0.7.2
Pillow==10.4.0
Flask==3.0.3
